requirements.txt
anyio==4.6.2.post1
blinker==1.6.2
certifi==2022.9.24
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
coverage==7.4.4
dryable==1.1.0
exceptiongroup==1.2.0
Flask==3.0.0
Flask-Cors==3.0.10
gunicorn==20.1.0
h11==0.14.0
halo==0.0.31
httpcore==1.0.6
httpx==0.27.2
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
log-symbols==0.0.14
MarkupSafe==2.1.2
notion-client==2.2.1
numpy==1.23.2
packaging==24.0
pg8000==1.12.5
pluggy==1.4.0
psycopg2-binary==2.9.7
pytest==8.1.1
pytest-cov==5.0.0
python-dotenv==0.12.0
pytz==2020.4
requests==2.27.1
sentry-sdk==1.34.0
six==1.16.0
sniffio==1.3.1
spinners==0.0.24
SQLAlchemy==1.3.15
termcolor==2.2.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==1.26.13
Werkzeug==3.0.1
zipp==3.17.0
